\name{regulatory_svm}
\alias{regulatory_svm}
\title{ Trains a SVM to recognize a certain pattern of regulatory positions }
\description{ Trains a SVM to recognize a certain pattern of regulatory positions.}
\usage{
regulatory_svm(gdm, 
      bw_plus_path, 
      bw_minus_path, 
      positions, positive, 
      allow = NULL, 
      n_train = 25000, 
      n_eval = 1000, 
      pdf_path = "roc_plot.pdf", 
      plot_raw_data = TRUE, 
      extra_enrich_bed = NULL, 
      extra_enrich_frac = 0.1, 
      enrich_negative_near_pos = 0.15, 
      use_rgtsvm = FALSE,
      svm_type = "SVR", 
      ncores = 1, 
      ..., 
      debug = TRUE)
}
\arguments{
  \item{gdm}{Genomic data model returned by \code{\link{genomic_data_model}}. }
  \item{bw_plus_path}{String indicating file path to bigWig file representing the plus strand.}
  \item{bw_minus_path}{String indicating file path to bigWig file representing the minus strand.}
  \item{positions}{Data frame with two columns(chrom,chromCenter), indicating the universe of positions to test and evaluate. It can be generated by \code{\link{get_informative_positions}}.}
  \item{positive}{Bed-style data frame containing positive positions(chrom,chromStart,chromEnd).}
  \item{allow}{Bed-style data frame containing positions to avoid in the negative set(chrom,chromStart,chromEnd).}
  \item{n_train}{Number of training examples.}
  \item{n_eval}{Number of examples on which to test performance.}
  \item{pdf_path}{String value indicating a PDF file.  Set to NULL if no PDF should be printed.}
  \item{plot_raw_data}{If TRUE (default), and if a PDF file is specified, plots the raw data used to train the model.}
  \item{extra_enrich_bed}{Bed-style data frame indicating extra bed file to enrich near. Used by \code{\link{get_test_set}}.}
  \item{extra_enrich_frac}{Fraction of final positions sampled in the negative set which are in the bed file. Unused if extra_enrich_bed is NULL. Used by \code{\link{get_test_set}}.}
  \item{enrich_negative_near_pos}{Fraction of training examples chosen to be nearby (<=5kb) a positive example [0,1].}
  \item{use_rgtsvm}{Indictating whether the predict will be performed on GPU through the Rgtsvm package.}
  \item{svm_type}{Two options, "SVR" for support vecctor regression (epsilon-regression). "P_SVM" for probabilistic SVM (C-classification).}
  \item{ncores}{Integer indicating how many cores are used to improve the performance.}
  \item{\dots}{The parameters for \code{plot} function}
  \item{debug}{Logical value indication the process detail is outputted.}
}

\value{
A svm model trained by \code{svm} function in e1071 package.
}
\references{
Danko, C. G., Hyland, S. L., Core, L. J., Martins, A. L., Waters, C. T., Lee, H. W., ... & Siepel, A. (2015). 
Identification of active transcriptional regulatory elements from GRO-seq data. Nature methods, 12(5), 433-438.
}
\seealso{
\code{\link{get_informative_positions}}, \code{\link{get_test_set}}, \code{\link{read_genomic_data}}, \code{\link{eval_reg_svm}}
}
\keyword{ function }
